refactor(cli): share app.yaml/databricks.yml binding contract between add and doctor
Extract the deploy-config binding shape into a shared module
(packages/shared/src/cli/deploy-config.ts): the databricks.yml/app.yaml
file-name constants, the AppYamlEnvEntry/ResourceBinding shapes, and the
inverse bindingToNode/bindingTypeOf pair.

'appkit add' (registry/config-writer) generates this deploy config and
'appkit doctor' (doctor/bundle) reads and validates it. Sharing the shape
plus the encode/decode pair keeps the writer from drifting off the reader.
No behavior change.
